Combine success with send application
