+++ app/assets/stylesheets/organisms/_form-card.scss
@@ -169,6 +169,7 @@
   }
 }
 
+.form-card__expect_next,
 .form-card__documents__faqs {
   @include full-bleed();
   @include clearfix();
