+++ app/views/documents/edit.html.erb
@@ -44,7 +44,7 @@
           Add documents
         </button>
 
-        <%= link_to send_application_steps_path, class: "button" do %>
+        <%= link_to next_path, class: "button" do %>
           I'll do this later
         <% end %>
       </div>
