--- app/views/send_application/edit.html.erb
@@ -1,28 +1,0 @@
-<% content_for :header_title, 'Email application' %>
-<div class="form-card">
-  <%= form_for @step, as: :step, builder: MbFormBuilder, url: current_path, method: :put do |f| %>
-    <header class="form-card__header">
-      <div class="step-section-header">
-        <div class="step-section-header__subhead">
-          Want a copy of your application?
-        </div>
-      </div>
-    </header>
-    <div class="form-card__content">
-      <p>
-        Enter your email address below to receive an email with a copy of your
-        application.
-      </p>
-      <%= f.mb_input_field :email, "Your email address", type: "email", options: { placeholder: "test@example.com" } %>
-    </div>
-    <footer class="form-card__footer">
-      <button class="button button--next button--cta button--full-width" type="submit">
-        Submit
-        <i class="button__icon icon-arrow_forward"></i>
-      </button>
-      <%= link_to skip_send_application_path, method: :post, class: "button" do %>
-        Skip this step
-      <% end %>
-    </footer>
-  <% end %>
-</div>
